Add persistence layer and implement delete product endpoint

+++ src/Controller/ProductController.php
@@ -3,6 +3,7 @@
 namespace App\Controller;
 
 use App\Entity\Product;
+use App\Repository\ProductRepository;
 use App\Request\ProductCreateRequest;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
@@ -18,21 +19,38 @@ final class ProductController extends AbstractController
     private SerializerInterface $serializer;
     private ValidatorInterface $validator;
     private EntityManagerInterface $entityManager;
+    private ProductRepository $productRepository;
 
-    public function __construct(SerializerInterface $serializer, ValidatorInterface $validator, EntityManagerInterface $entityManager)
-    {
+    public function __construct(
+        SerializerInterface $serializer,
+        ValidatorInterface $validator,
+        ProductRepository $productRepository,
+        EntityManagerInterface $entityManager
+    ) {
         $this->serializer = $serializer;
         $this->validator = $validator;
+        $this->productRepository = $productRepository;
         $this->entityManager = $entityManager;
     }
 
     #[Route('/api/products/{id}', name: 'app_product_delete', methods: ['DELETE'])]
-    public function delete(Product $product): JsonResponse
+    public function delete(int $id): JsonResponse
     {
+        $product = $this->productRepository->find($id);
+
+        if (!$product) {
+            return new JsonResponse([
+                'message' => 'Product not found.'
+            ],
+                Response::HTTP_NOT_FOUND,
+            [
+                'Content-Type' => 'application/json; charset=utf-8'
+            ]);
+        }
         $this->entityManager->remove($product);
         $this->entityManager->flush();
 
-        return new JsonResponse([],Response::HTTP_NO_CONTENT);
+        return new JsonResponse(null, Response::HTTP_NO_CONTENT);
     }
 
     #[Route('/api/products', name: 'app_product_create', methods: ['POST'])]
@@ -58,14 +76,20 @@ final class ProductController extends AbstractController
             ], Response::HTTP_UNPROCESSABLE_ENTITY);
         }
 
+        $product = new Product();
+        $product->setTitle($productCreateRequest->title);
+        $product->setPrice($productCreateRequest->price);
+        $this->entityManager->persist($product);
+        $this->entityManager->flush();
+
         return new JsonResponse([
-            'id' => 1,
-            'title' => $productCreateRequest->title,
-            'price' => $productCreateRequest->price,
+            'id' => $product->getId(),
+            'title' => $product->getTitle(),
+            'price' => $product->getPrice(),
         ],
             Response::HTTP_CREATED,
-            [
-                'Content-Type' => 'application/json; charset=utf-8',
+        [
+            'Content-Type' => 'application/json; charset=utf-8',
         ]);
     }
 }
